Azure Functions/DemoHttpTrigger/requirements.txt
 
azure-functions 
azure-identity 
azure-storage-blob 
pandas 
python-dotenv 
pydantic 
azure-monitor-opentelemetry 
azure-monitor-events-extension 
